requirements.txt
langgraph
langchain-chroma
langchain-community
langchain[openai]
langchain-text-splitters
pymupdf
chromadb
python-dotenv
pandas
tiktoken